Fixed unnamed function pointers
This just requires a special Expr::Splat type for standalone splats,
similar to Expr::DotDotDot for ... in functions with varargs.

We already have parsing for Suffnary splats, so I think this was just an
oversight.

---

I think this also fixes the splat mess from earlier, though I'm not sure
exactly what the problem was... Nested splats should be handled by the
subexpr parser?